incorrect handling of script arguments with spaces
Running `examples/testfile.nim klm "abc def"` will result in the last line of output being `@["klm", "abc", "def"]` which is incorrect. As 
it should be `@["klm", "abc def"]`.

The cause for this is the incorrect stringification, without quotes, for `execShellCmd`.  Using `startProcess` you can hand in the already split arguments, and not have the overhead of another shell invocation.

Fixes issue #3